requirements.txt
PyYAML==6.0
Click==7.0
rich==12.6.0
